requirements.txt
RPi.GPIO
requests
pyyaml